+++ DTCD-FooterPanel/src/Plugin.js
@@ -21,7 +21,5 @@ export class FooterPanel extends PanelPlugin {
       data: () => data,
       render: h => h(PluginComponent),
     }).$mount(selector);
-
-    styleSystem.setVariablesToElement(vue.$el, styleSystem.getCurrentTheme());
   }
 }